requirements.txt
tensorflow>=2.0.0
streamlit>=1.0.0
plotly>=4.14.3
pandas>=1.2.0
numpy>=1.19.2
Pillow>=8.0.0
scikit-learn>=0.24.0
seaborn>=0.11.0
matplotlib>=3.3.0
python-dotenv>=0.19.0
tqdm>=4.62.0